Split module incremental
