+++ src/engine/incremental/mod.rs
@@ -1,11 +1,8 @@
-mod cmd_outputs;
-mod fs;
+mod env_state;
 
-use crate::domain::{EnvProbes, Target};
+use crate::domain::Target;
 use anyhow::{Context, Error, Result};
-use cmd_outputs::EnvCmdOutputsState;
-use fs::EnvFsState;
-use serde::{Deserialize, Serialize};
+use env_state::TargetEnvState;
 use std::fs::File;
 use std::io::ErrorKind;
 use std::path::{self, Path, PathBuf};
@@ -145,53 +142,3 @@ pub fn remove_checksums_dir(project_dir: PathBuf) -> Result<()> {
 
     Ok(())
 }
-
-#[derive(Serialize, Deserialize, PartialEq)]
-struct TargetEnvState {
-    inputs: EnvState,
-    outputs: EnvState,
-}
-
-impl TargetEnvState {
-    fn current(target: &Target) -> Result<Option<Self>> {
-        if target.inputs.is_empty() {
-            Ok(None)
-        } else {
-            let project_dir = &target.project.dir;
-            Ok(Some(TargetEnvState {
-                inputs: EnvState::current(&target.inputs, project_dir)?,
-                outputs: EnvState::current(&target.outputs, project_dir)?,
-            }))
-        }
-    }
-
-    fn eq_current_state(&self, target: &Target) -> Result<bool> {
-        let project_dir = &target.project.dir;
-        Ok(self.inputs.eq_current_state(&target.inputs, &project_dir)?
-            && self
-                .outputs
-                .eq_current_state(&target.outputs, &project_dir)?)
-    }
-}
-
-#[derive(Serialize, Deserialize, PartialEq)]
-struct EnvState {
-    fs: EnvFsState,
-    cmd_stdouts: EnvCmdOutputsState,
-}
-
-impl EnvState {
-    fn current(env_probes: &EnvProbes, project_dir: &Path) -> Result<Self> {
-        Ok(Self {
-            fs: EnvFsState::current(&env_probes.paths)?,
-            cmd_stdouts: EnvCmdOutputsState::current(&env_probes.cmd_outputs, project_dir)?,
-        })
-    }
-
-    fn eq_current_state(&self, env_probes: &EnvProbes, project_dir: &Path) -> Result<bool> {
-        Ok((&self.fs).eq_current_state(&env_probes.paths)?
-            && (&self.cmd_stdouts).eq_current_state(&env_probes.cmd_outputs, project_dir)?)
-    }
-}
-
-// TODO Run all computations in parallel?
